Update en.php
Changed the error messages for min and max. If I set a min value of 1 the error message says the value has to be greater than 1 - changed the error message to say the value must be at least 1. If I set a max value of 10 the error message says the value has to be less than 10 - changed the error message to say the value must be no more than 10.