update language file to improve error messages
Correct me if I am wrong but the length validation checks that the value is the same length as the given parameter, neither more nor less. I have corrected the message to reflect this.

Also adjusted the lengthMin and lengthMax messages to reflect the fact that you can match the given parameter but not exceed or have less than it.